Move to rum-core and reuse in recorder
