+++ packages/core/src/tools/utils/stringUtils.spec.ts
@@ -1,4 +1,4 @@
-import { safeTruncate, findCommaSeparatedValue, findCommaSeparatedValues, findDataUrlAndTruncate } from './stringUtils'
+import { safeTruncate, findCommaSeparatedValue, findCommaSeparatedValues } from './stringUtils'
 
 describe('stringUtils', () => {
   describe('safeTruncate', () => {
@@ -55,36 +55,6 @@ describe('stringUtils', () => {
       expectedValues.set('foo', 'a')
       expectedValues.set('bar', 'b')
       expect(findCommaSeparatedValues('foo=a;bar=b')).toEqual(expectedValues)
-    })
-  })
-
-  fdescribe('findDataUrlAndTruncate', () => {
-    it('returns truncated url when detects data url of json', () => {
-      const expectedUrl = 'data:text/json'
-      expect(
-        findDataUrlAndTruncate(
-          'data:text/json; charset=utf-8,%7B%22data%22%3A%7B%22type%22%3A%22notebooks%22%2C%22attributes%22%3A%7B%22metadata%22%3A%7B'
-        )
-      ).toEqual(expectedUrl)
-    })
-
-    it('returns truncated url when detects data url of html', () => {
-      const expectedUrl = 'data:text/html'
-      expect(findDataUrlAndTruncate('data:text/html,%3Ch1%3EHello%2C%20World%21%3C%2Fh1%3E')).toEqual(expectedUrl)
-    })
-
-    it('returns truncated url when detects data url of image', () => {
-      const expectedUrl = 'data:image/svg+xml;base64'
-      expect(findDataUrlAndTruncate('data:image/svg+xml;base64,+DQo8L3N2Zz4=')).toEqual(expectedUrl)
-    })
-    it('returns truncated url when detects plain data url', () => {
-      const expectedUrl = 'data:'
-      expect(findDataUrlAndTruncate('data:,Hello%2C%20World%21')).toEqual(expectedUrl)
-    })
-
-    it('returns null when no data url found', () => {
-      const nonDataUrl = 'https://static.datad0g.com/static/c/70086/chunk.min.js'
-      expect(findDataUrlAndTruncate(nonDataUrl)).toBeNull()
     })
   })
 })
