+++ packages/rum-core/src/domain/resource/resourceUtils.spec.ts
@@ -8,6 +8,7 @@ import {
   computePerformanceResourceDuration,
   computeResourceKind,
   isAllowedRequestUrl,
+  findDataUrlAndTruncate,
 } from './resourceUtils'
 
 function generateResourceWith(overrides: Partial<RumPerformanceResourceTiming>) {
@@ -311,5 +312,35 @@ describe('shouldTrackResource', () => {
 
   it('should allow requests on non intake domains', () => {
     expect(isAllowedRequestUrl(configuration, 'https://my-domain.com/hello?a=b')).toBe(true)
+  })
+})
+
+describe('findDataUrlAndTruncate', () => {
+  it('returns truncated url when detects data url of json', () => {
+    const expectedUrl = 'data:text/json'
+    expect(
+      findDataUrlAndTruncate(
+        'data:text/json; charset=utf-8,%7B%22data%22%3A%7B%22type%22%3A%22notebooks%22%2C%22attributes%22%3A%7B%22metadata%22%3A%7B'
+      )
+    ).toEqual(expectedUrl)
+  })
+
+  it('returns truncated url when detects data url of html', () => {
+    const expectedUrl = 'data:text/html'
+    expect(findDataUrlAndTruncate('data:text/html,%3Ch1%3EHello%2C%20World%21%3C%2Fh1%3E')).toEqual(expectedUrl)
+  })
+
+  it('returns truncated url when detects data url of image', () => {
+    const expectedUrl = 'data:image/svg+xml;base64'
+    expect(findDataUrlAndTruncate('data:image/svg+xml;base64,+DQo8L3N2Zz4=')).toEqual(expectedUrl)
+  })
+  it('returns truncated url when detects plain data url', () => {
+    const expectedUrl = 'data:'
+    expect(findDataUrlAndTruncate('data:,Hello%2C%20World%21')).toEqual(expectedUrl)
+  })
+
+  it('returns null when no data url found', () => {
+    const nonDataUrl = 'https://static.datad0g.com/static/c/70086/chunk.min.js'
+    expect(findDataUrlAndTruncate(nonDataUrl)).toBeNull()
   })
 })
